refact : model 정의 분리 app.js model 정의 부분 제거 후 models/ 폴더로 분리
