+++ models/Posts.js
@@ -1,0 +1,20 @@
+const { DataTypes } = require('sequelize');
+const sequelize = require('../conf/database');
+
+const Post = sequelize.define('posts', {
+    id: {
+        type: DataTypes.INTEGER,
+        primaryKey: true,
+        autoIncrement: true,
+    },
+    title: {
+        type: DataTypes.STRING,
+        allowNull: false,
+    },
+    content: {
+        type: DataTypes.STRING,
+        allowNull: false,
+    }
+})
+
+module.exports = Post;
