+++ models/index.js
@@ -1,0 +1,10 @@
+const User = require('./Users');
+const Post = require('./Posts');
+
+User.hasMany(Post);
+Post.belongsTo(User);
+
+module.exports = {
+    User,
+    Post
+}
